requirements.txt
flask==2.3.3
google-api-python-client==2.108.0
pandas==2.1.3
scikit-learn==1.3.2
numpy==1.24.3
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
